StinkyTofu.Core.Hints

Documentation

data Hints

Constructors

HZero

HUnit

HIf

HFold

HNot

HShl1

HShr1

HShr4

HShr16

HAnd

HOr

HXor

HPlus

HVar VName

consume :: Hints -> Int

consume return the consumming amount of a Hints, i.e. consume = 1 + addFork.

addFork :: Hints -> Int

addFork the amount of additional branches an operator will provided.